Fix link-type check warnings on kotlin-stdlib
This change fixes the link-type check warning from an app built with SDK
to the prebuilt module kotlin-stdlib. The warnings were triggered since
the prebuilt module is declared without the sdk_version property and thus
it is incorrectly identified as a lib using private APIs.

Fixing the warning simply by setting sdk_version property to
"core_current". The value was chosen instead of "current" because the
kotlin library does not have any dependency to Android-specific APIs.

Bug: 69899800
Test: m -j android-arch-room-migration is successful and does not show
any link-type warning (for now the test must be run in the internal
master as dependency to kotlin-stdlib is not yet on AOSP)

Change-Id: I66c296bea3a92540a8a0546fb824a8aa2118857e