Add kotlin-test and kotlin-stdlib-renamed libraries
kotlin-test is backed by junit and allows to write kotlin tests.
kotlin-stdlib-renamed is a kotlin standard library that can be used in
framework/platform android code. It's renamed to com.android.kotlin.*

kotlin.test is not renambed due to its ServiceLoader usage
to load AsserterContributor implementation. ServiceLoader is
not happy with jarjar renaming.

Bug: 73281388
Test: tbd
Change-Id: I7f58062ac9a377f0bcf800df068bf94d83110d24